support debug mode

+++ test/plugin/containers/jvm-container/docker/run.sh
@@ -16,13 +16,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 function exitOnError() {
     echo -e "\033[31m[ERROR] $1\033[0m">&2
     exitAndClean 1
 }
 
 function exitAndClean() {
+    [[ -z $DEBUG_MODE ]] && exit $1;
+
     [[ -f ${SCENARIO_HOME}/data/actualData.yaml ]] && rm -rf ${SCENARIO_HOME}/data/actualData.yaml
     [[ -d ${SCENARIO_HOME}/logs ]] && rm -rf ${SCENARIO_HOME}/logs
     exit $1
@@ -53,13 +54,15 @@ TOOLS_HOME=/usr/local/skywalking/tools
 SCENARIO_HOME=/usr/local/skywalking/scenario
 JACOCO_HOME=${JACOCO_HOME:-/jacoco}
 
+mkdi -p ${LOGS_HOME}
+
 unzip -q ${SCENARIO_HOME}/*.zip -d /var/run/
 if [[ ! -f /var/run/${SCENARIO_NAME}/${SCENARIO_START_SCRIPT} ]]; then
     exitOnError "The required startup script not exists!"
 fi
 
 echo "To start mock collector"
-${TOOLS_HOME}/skywalking-mock-collector/bin/collector-startup.sh 1>/dev/null &
+${TOOLS_HOME}/skywalking-mock-collector/bin/collector-startup.sh 1>${SCENARIO_HOME}/logs/collector.out &
 healthCheck http://localhost:12800/receiveData
 
 # start applications
@@ -71,10 +74,10 @@ export agent_opts="
     -Dskywalking.collector.discovery_check_interval=2
     -Dskywalking.collector.backend_service=localhost:19876
     -Dskywalking.agent.service_name=${SCENARIO_NAME}
-    -Dskywalking.logging.dir=/usr/local/skywalking/scenario/logs
+    -Dskywalking.logging.dir=${SCENARIO_HOME}/logs
     -Dskywalking.agent.authentication=test-token
     -Xms256m -Xmx256m ${agent_opts}"
-exec /var/run/${SCENARIO_NAME}/${SCENARIO_START_SCRIPT} 1>/dev/null &
+exec /var/run/${SCENARIO_NAME}/${SCENARIO_START_SCRIPT} 1>${SCENARIO_HOME}/logs/scenario.out &
 
 healthCheck ${SCENARIO_HEALTH_CHECK_URL}
 
@@ -91,7 +94,7 @@ java -jar \
     -Xmx256m -Xms256m \
     -DcaseName="${SCENARIO_NAME}-${SCENARIO_VERSION}" \
     -DtestCasePath=${SCENARIO_HOME}/data/ \
-    ${TOOLS_HOME}/skywalking-validator-tools.jar 1>/dev/null
+    ${TOOLS_HOME}/skywalking-validator-tools.jar 1>1>${SCENARIO_HOME}/logs/validator.out
 status=$?
 
 if [[ $status -eq 0 ]]; then
